+++ pwchem/protocols/Sequences/protocol_define_sequence_roi.py
