+++ pwchem/protocols/Sequences/protocol_import_setOfSequences.py
