+++ pwchem/protocols/Sequences/protocol_operate_sequence_rois.py
