+++ pwchem/protocols/VirtualDrugScreening/protocol_define_manual_pockets.py
